chore: created a directory for headers. parser ast print non-functioning
dont know why the printer doesnt work. think the build system is broken

added header for ast ops

renamed parser.y -> grammar.y

purged some old files for testing

+++ include/lexer/lexer.lex.h
@@ -32,7 +32,7 @@ union
     char *s;
 };
 
-size_t s_len;   // Only used for strings
+int s_len;   // Only used for strings
 unsigned int tags;
 /*  bit meaning (1 is true)
     0   unsigned
